[gardening] Silence compiler warnings [NFC]

+++ stdlib/public/core/UTF16.swift
@@ -41,10 +41,10 @@ extension Unicode.UTF16 : Unicode.Encoding {
     _ source: Unicode.Scalar
   ) -> EncodedScalar? {
     let x = source.value
-    if _fastPath(x < (1 << 16)) {
+    if _fastPath(x < ((1 as UInt32) << 16)) {
       return EncodedScalar(_storage: x, _bitCount: 16)
     }
-    let x1 = x - (1 << 16)
+    let x1 = x - ((1 as UInt32) << 16)
     var r = (0xdc00 + (x1 & 0x3ff))
     r &<<= 16
     r |= (0xd800 + (x1 &>> 10 & 0x3ff))
